fix: deprecating legacy id token auth
fix: fix issuer in testing executor

fix: testing package pnpm-lock

fix: fixing things and prettier

fix: cleaning up
