+++ runtime/actions/authenticate.go
@@ -8,14 +8,10 @@ import (
 	email "net/mail"
 	"net/url"
 	"strings"
-	"time"
 
-	"github.com/golang-jwt/jwt/v4"
 	"github.com/karlseguin/typed"
-	"github.com/samber/lo"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/codes"
-	"go.opentelemetry.io/otel/trace"
 
 	"github.com/teamkeel/keel/mail"
 	"github.com/teamkeel/keel/proto"
@@ -40,24 +36,7 @@ var (
 	ErrIdentityNotFound = common.NewAuthenticationFailedMessageErr("identity not found")
 )
 
-const (
-	DefaultBearerTokenExpiry time.Duration = time.Hour * 24
-	ResetTokenExpiry         time.Duration = time.Minute * 15
-)
-
-const (
-	resetPasswordAudClaim = "password-reset"
-	keelIssuerClaim       = "keel"
-)
-
-// https://pkg.go.dev/github.com/golang-jwt/jwt/v4#RegisteredClaims
-type Claims struct {
-	jwt.RegisteredClaims
-}
-
 // Authenticate will return the identity ID if it is successfully authenticated or when a new identity is created.
-//
-// Deprecated: we will be deprecating the authenticate action and password flow in favour of the new auth endpoints
 func Authenticate(scope *Scope, input map[string]any) (*AuthenticateResult, error) {
 	typedInput := typed.New(input)
 
@@ -70,7 +49,7 @@ func Authenticate(scope *Scope, input map[string]any) (*AuthenticateResult, erro
 		return nil, common.RuntimeError{Code: common.ErrInvalidInput, Message: "password cannot be empty"}
 	}
 
-	identity, err := FindIdentityByEmail(scope.Context, scope.Schema, emailPassword.String("email"), keelIssuerClaim)
+	identity, err := FindIdentityByEmail(scope.Context, scope.Schema, emailPassword.String("email"), oauth.KeelIssuer)
 	if err != nil {
 		return nil, err
 	}
@@ -81,7 +60,7 @@ func Authenticate(scope *Scope, input map[string]any) (*AuthenticateResult, erro
 			return nil, common.RuntimeError{Code: common.ErrInvalidInput, Message: "failed to authenticate"}
 		}
 
-		token, err := GenerateBearerToken(scope.Context, identity.Id)
+		token, _, err := oauth.GenerateAccessToken(scope.Context, identity.Id)
 		if err != nil {
 			return nil, err
 		}
@@ -96,25 +75,17 @@ func Authenticate(scope *Scope, input map[string]any) (*AuthenticateResult, erro
 		return nil, common.RuntimeError{Code: common.ErrInvalidInput, Message: "failed to authenticate"}
 	}
 
-	config, err := runtimectx.GetAuthConfig(scope.Context)
-	if err == nil {
-		if config != nil && config.Keel != nil && !config.Keel.AllowCreate {
-			// Creating new identities is disabled
-			return nil, common.RuntimeError{Code: common.ErrInvalidInput, Message: "registration disabled"}
-		}
-	}
-
 	hashedBytes, err := bcrypt.GenerateFromPassword([]byte(emailPassword.String("password")), bcrypt.DefaultCost)
 	if err != nil {
 		return nil, err
 	}
 
-	identity, err = CreateIdentity(scope.Context, scope.Schema, emailPassword.String("email"), string(hashedBytes), keelIssuerClaim)
+	identity, err = CreateIdentity(scope.Context, scope.Schema, emailPassword.String("email"), string(hashedBytes), oauth.KeelIssuer)
 	if err != nil {
 		return nil, err
 	}
 
-	token, err := GenerateBearerToken(scope.Context, identity.Id)
+	token, _, err := oauth.GenerateAccessToken(scope.Context, identity.Id)
 	if err != nil {
 		return nil, err
 	}
@@ -125,7 +96,6 @@ func Authenticate(scope *Scope, input map[string]any) (*AuthenticateResult, erro
 	}, nil
 }
 
-// Deprecated: we will be deprecating the authenticate action and password flow in favour of the new auth endpoints
 func ResetRequestPassword(scope *Scope, input map[string]any) error {
 	var err error
 	typedInput := typed.New(input)
@@ -141,7 +111,7 @@ func ResetRequestPassword(scope *Scope, input map[string]any) error {
 	}
 
 	var identity *auth.Identity
-	identity, err = FindIdentityByEmail(scope.Context, scope.Schema, emailString, keelIssuerClaim)
+	identity, err = FindIdentityByEmail(scope.Context, scope.Schema, emailString, oauth.KeelIssuer)
 	if err != nil {
 		return err
 	}
@@ -149,7 +119,7 @@ func ResetRequestPassword(scope *Scope, input map[string]any) error {
 		return nil
 	}
 
-	token, err := GenerateResetToken(scope.Context, identity.Id)
+	token, err := oauth.GenerateResetToken(scope.Context, identity.Id)
 	if err != nil {
 		return err
 	}
@@ -180,7 +150,7 @@ func ResetPassword(scope *Scope, input map[string]any) error {
 	token := typedInput.String("token")
 	password := typedInput.String("password")
 
-	identityId, err := ValidateResetToken(scope.Context, token)
+	identityId, err := oauth.ValidateResetToken(scope.Context, token)
 	switch {
 	case errors.Is(err, ErrInvalidToken) || errors.Is(err, ErrTokenExpired):
 		return common.RuntimeError{Code: common.ErrInvalidInput, Message: err.Error()}
@@ -214,209 +184,6 @@ func ResetPassword(scope *Scope, input map[string]any) error {
 	return nil
 }
 
-func GenerateBearerToken(ctx context.Context, identityId string) (string, error) {
-	expiry := DefaultBearerTokenExpiry
-	config, err := runtimectx.GetAuthConfig(ctx)
-	if err == nil {
-		if config != nil && config.Keel != nil {
-			expiry = time.Duration(config.Keel.TokenDuration) * time.Second
-		}
-	}
-
-	return generateToken(ctx, identityId, []string{}, expiry)
-}
-
-// Deprecated: we will be deprecating the authenticate action and password flow in favour of the new auth endpoints
-func GenerateResetToken(ctx context.Context, identityId string) (string, error) {
-	return generateToken(ctx, identityId, []string{resetPasswordAudClaim}, ResetTokenExpiry)
-}
-
-func generateToken(ctx context.Context, sub string, aud []string, expiresIn time.Duration) (string, error) {
-	now := time.Now().UTC()
-	claims := Claims{
-		RegisteredClaims: jwt.RegisteredClaims{
-			Subject:   sub,
-			Audience:  aud,
-			ExpiresAt: jwt.NewNumericDate(now.Add(expiresIn)),
-			IssuedAt:  jwt.NewNumericDate(now),
-			Issuer:    keelIssuerClaim,
-		},
-	}
-
-	privateKey, err := runtimectx.GetPrivateKey(ctx)
-	if err != nil {
-		return "", err
-	}
-
-	if privateKey == nil {
-		return "", fmt.Errorf("no private key set")
-	}
-
-	token := jwt.NewWithClaims(jwt.SigningMethodRS256, claims)
-	tokenString, err := token.SignedString(privateKey)
-	if err != nil {
-		return "", fmt.Errorf("cannot create signed jwt: %w", err)
-	}
-	return tokenString, nil
-}
-
-// Verifies the bearer token and returns the JWT subject and issuer.
-//
-// Deprecated: we will be deprecating the authenticate action and password flow in favour of the new auth endpoints
-func ValidateBearerToken(ctx context.Context, tokenString string) (string, string, error) {
-	return validateToken(ctx, tokenString, "")
-}
-
-// Verifies the reset token and returns the JWT subject.
-//
-// Deprecated: we will be deprecating the authenticate action and password flow in favour of the new auth endpoints
-func ValidateResetToken(ctx context.Context, tokenString string) (string, error) {
-	subject, issuer, err := validateToken(ctx, tokenString, resetPasswordAudClaim)
-	if issuer != keelIssuerClaim && issuer != "" {
-		return "", fmt.Errorf("can only accept reset token from %s issuer, not: %s", keelIssuerClaim, issuer)
-	}
-	return subject, err
-}
-
-// Deprecated: we will be deprecating the authenticate action and password flow in favour of the new auth endpoints
-func validateToken(ctx context.Context, tokenString string, audienceClaim string) (string, string, error) {
-	privateKey, err := runtimectx.GetPrivateKey(ctx)
-	if err != nil {
-		return "", "", err
-	}
-
-	if privateKey == nil {
-		return "", "", errors.New("no private key set")
-	}
-
-	ctx, span := tracer.Start(ctx, "Validate token")
-	defer span.End()
-
-	authConfig, _ := runtimectx.GetAuthConfig(ctx)
-
-	var token *jwt.Token
-	claims := &Claims{}
-
-	// try to decode the token and validate using our private key as the signing method.
-	// this supports external issued tokens but which are signed with our private key (such as clerk)
-	span.AddEvent("Validating with Keel pk")
-	token, err = jwt.ParseWithClaims(tokenString, claims, func(t *jwt.Token) (interface{}, error) {
-		return &privateKey.PublicKey, nil
-	})
-
-	var validationErr *jwt.ValidationError
-	if errors.As(err, &validationErr) && validationErr.Errors == jwt.ValidationErrorExpired {
-		return "", "", ErrTokenExpired
-	}
-
-	// if unsuccessful using our private key, try to validate against any of the known external issuers if there are any
-	if err != nil {
-		token, err = jwt.ParseWithClaims(tokenString, claims, func(t *jwt.Token) (interface{}, error) {
-			iss := t.Claims.(*Claims).Issuer
-
-			if authConfig == nil {
-				// If no auth config skip all this
-				return nil, nil
-			}
-
-			issuers := authConfig.Issuers
-
-			if authConfig.AllowAnyIssuers {
-				span.AddEvent("Accepting any issuer as AllowAnyIssuers is enabled")
-				// In this mode we allow any issuer that is openID connect compatible
-				// So if this issuer is new add it to the know issuers and verify
-
-				match := false
-				for _, extIssuers := range issuers {
-					if extIssuers.Iss == iss {
-						match = true
-					}
-				}
-				if !match {
-					if iss != "" {
-						issuers = append(issuers, runtimectx.ExternalIssuer{
-							Iss: iss,
-						})
-					}
-				}
-
-				authConfig.Issuers = issuers
-				ctx = runtimectx.WithAuthConfig(ctx, *authConfig)
-
-			}
-
-			if len(issuers) > 0 {
-				span.AddEvent("Validating with external issuers")
-			}
-
-			for _, issuer := range issuers {
-				if issuer.Iss != iss {
-					continue
-				}
-				iss := t.Claims.(*Claims).Issuer
-
-				kid := ""
-				if header, ok := t.Header["kid"]; ok {
-					if kidStr, ok := header.(string); ok {
-						kid = kidStr
-					}
-				}
-
-				publicKey, err := auth.PublicKeyForIssuer(ctx, iss, kid)
-
-				// Check the audience matches if set
-				match := true
-				if issuer.Audience != nil {
-					match = false
-					aud := t.Claims.(*Claims).Audience
-					for _, audience := range aud {
-						if audience == *issuer.Audience {
-							match = true
-							break
-						}
-					}
-					span.AddEvent("Validating audience claims", trace.WithAttributes(attribute.Bool("match", match)))
-				}
-
-				if !match {
-					return nil, fmt.Errorf("invalid token")
-				}
-
-				if err == nil {
-					return publicKey, nil
-				}
-			}
-
-			return nil, fmt.Errorf("unexpected issuer in token: %s", iss)
-		})
-	}
-
-	if err != nil {
-		return "", "", ErrInvalidToken
-	}
-
-	if !claims.VerifyExpiresAt(time.Now().UTC(), true) {
-		return "", "", ErrTokenExpired
-	}
-
-	if audienceClaim != "" {
-		if !lo.Contains(claims.Audience, audienceClaim) {
-			return "", "", ErrInvalidToken
-		}
-	}
-
-	if err != nil || !token.Valid {
-		return "", "", ErrInvalidToken
-	}
-
-	if claims.Subject == "" {
-		return "", "", errors.New("subject claim cannot be empty")
-	}
-
-	return claims.Subject, claims.Issuer, nil
-}
-
-// Deprecated: we will be deprecating the authenticate action and password flow in favour of the new auth endpoints
 func HandleAuthorizationHeader(ctx context.Context, schema *proto.Schema, headers http.Header) (*auth.Identity, error) {
 	header := headers.Get("Authorization")
 	if header == "" {
@@ -441,29 +208,17 @@ func HandleAuthorizationHeader(ctx context.Context, schema *proto.Schema, header
 	return nil, nil
 }
 
-// Deprecated: we will be deprecating the authenticate action and password flow in favour of the new auth endpoints
 func HandleBearerToken(ctx context.Context, schema *proto.Schema, token string) (*auth.Identity, error) {
 	ctx, span := tracer.Start(ctx, "Authorization")
 	defer span.End()
 
-	subject, issuer, err := ValidateBearerToken(ctx, token)
+	subject, err := oauth.ValidateAccessToken(ctx, token)
 	if err != nil {
 		span.SetStatus(codes.Error, err.Error())
 		return nil, err
 	}
 
-	// Check that identity actually does exist as it could
-	// have been deleted after the bearer token was generated.
-	var identity *auth.Identity
-	if issuer == "keel" || issuer == "" || issuer == oauth.KeelIssuer {
-		identity, err = FindIdentityById(ctx, schema, subject)
-	} else {
-		identity, err = FindIdentityByExternalId(ctx, schema, subject, issuer)
-		if identity == nil {
-			identity, err = CreateExternalIdentity(ctx, schema, subject, issuer, token)
-		}
-	}
-
+	identity, err := FindIdentityById(ctx, schema, subject)
 	if err != nil {
 		return nil, err
 	}
