+++ runtime/oauth/access_token_test.go
@@ -34,10 +34,9 @@ func TestAccessTokenGeneration(t *testing.T) {
 	require.NoError(t, err)
 	require.NotEmpty(t, bearerJwt)
 
-	parsedId, iss, err := oauth.ValidateAccessToken(ctx, bearerJwt)
+	parsedId, err := oauth.ValidateAccessToken(ctx, bearerJwt)
 	require.NoError(t, err)
 	require.Equal(t, identityId.String(), parsedId)
-	require.Equal(t, oauth.KeelIssuer, iss)
 }
 
 func TestAccessTokenValidationNoPrivateKey(t *testing.T) {
@@ -50,10 +49,9 @@ func TestAccessTokenValidationNoPrivateKey(t *testing.T) {
 
 	ctx = runtimectx.WithPrivateKey(ctx, nil)
 
-	parsedId, iss, err := oauth.ValidateAccessToken(ctx, bearerJwt)
+	parsedId, err := oauth.ValidateAccessToken(ctx, bearerJwt)
 	require.Error(t, err, "no private key set")
 	require.Empty(t, parsedId)
-	require.Empty(t, iss)
 }
 
 func TestAccessTokenGenerationAndParsingWithSamePrivateKey(t *testing.T) {
@@ -64,10 +62,9 @@ func TestAccessTokenGenerationAndParsingWithSamePrivateKey(t *testing.T) {
 	require.NoError(t, err)
 	require.NotEmpty(t, bearerJwt)
 
-	parsedId, iss, err := oauth.ValidateAccessToken(ctx, bearerJwt)
+	parsedId, err := oauth.ValidateAccessToken(ctx, bearerJwt)
 	require.NoError(t, err)
 	require.Equal(t, identityId.String(), parsedId)
-	require.Equal(t, oauth.KeelIssuer, iss)
 }
 
 func TestAccessTokenValidationDifferentPrivateKey(t *testing.T) {
@@ -84,10 +81,9 @@ func TestAccessTokenValidationDifferentPrivateKey(t *testing.T) {
 	require.NotEmpty(t, bearerJwt)
 
 	ctx = newContextWithPK()
-	parsedId, iss, err := oauth.ValidateAccessToken(ctx, bearerJwt)
+	parsedId, err := oauth.ValidateAccessToken(ctx, bearerJwt)
 	require.ErrorIs(t, oauth.ErrInvalidToken, err)
 	require.Empty(t, parsedId)
-	require.Empty(t, iss)
 }
 
 func TestAccessTokenIsRSAMethodWithPrivateKey(t *testing.T) {
@@ -169,10 +165,9 @@ func TestShortExpiredAccessTokenIsInvalid(t *testing.T) {
 
 	time.Sleep(1100 * time.Millisecond)
 
-	parsedId, iss, err := oauth.ValidateAccessToken(ctx, bearerJwt)
+	parsedId, err := oauth.ValidateAccessToken(ctx, bearerJwt)
 	require.ErrorIs(t, oauth.ErrTokenExpired, err)
 	require.Empty(t, parsedId)
-	require.Empty(t, iss)
 }
 
 func TestExpiredAccessTokenIsInvalid(t *testing.T) {
@@ -198,8 +193,167 @@ func TestExpiredAccessTokenIsInvalid(t *testing.T) {
 	tokenString, err := token.SignedString(privateKey)
 	require.NoError(t, err)
 
-	parsedId, iss, err := oauth.ValidateAccessToken(ctx, tokenString)
+	parsedId, err := oauth.ValidateAccessToken(ctx, tokenString)
 	require.ErrorIs(t, oauth.ErrTokenExpired, err)
 	require.Empty(t, parsedId)
-	require.Empty(t, iss)
+}
+
+func TestResetTokenGenerationAndParsingWithoutPrivateKey(t *testing.T) {
+	ctx := newContextWithPK()
+	ctx = runtimectx.WithEnv(ctx, runtimectx.KeelEnvTest)
+	identityId := ksuid.New()
+
+	bearerJwt, err := oauth.GenerateResetToken(ctx, identityId.String())
+	require.NoError(t, err)
+	require.NotEmpty(t, bearerJwt)
+
+	parsedId, err := oauth.ValidateResetToken(ctx, bearerJwt)
+	require.NoError(t, err)
+	require.Equal(t, identityId.String(), parsedId)
+}
+
+func TestResetTokenGenerationAndParsingWithSamePrivateKey(t *testing.T) {
+	ctx := newContextWithPK()
+	identityId := ksuid.New()
+
+	privateKey, err := rsa.GenerateKey(rand.Reader, 2048)
+	require.NoError(t, err)
+	ctx = runtimectx.WithPrivateKey(ctx, privateKey)
+	require.NoError(t, err)
+
+	bearerJwt, err := oauth.GenerateResetToken(ctx, identityId.String())
+	require.NoError(t, err)
+	require.NotEmpty(t, bearerJwt)
+
+	parsedId, err := oauth.ValidateResetToken(ctx, bearerJwt)
+	require.NoError(t, err)
+	require.Equal(t, identityId.String(), parsedId)
+}
+
+func TestResetTokenGenerationAndParsingWithDifferentPrivateKeys(t *testing.T) {
+	ctx := newContextWithPK()
+	identityId := ksuid.New()
+
+	privateKey1, err := rsa.GenerateKey(rand.Reader, 2048)
+	require.NoError(t, err)
+	ctx = runtimectx.WithPrivateKey(ctx, privateKey1)
+	require.NoError(t, err)
+
+	bearerJwt, err := oauth.GenerateResetToken(ctx, identityId.String())
+	require.NoError(t, err)
+	require.NotEmpty(t, bearerJwt)
+
+	privateKey2, err := rsa.GenerateKey(rand.Reader, 2048)
+	require.NoError(t, err)
+	ctx = runtimectx.WithPrivateKey(ctx, privateKey2)
+	require.NoError(t, err)
+
+	parsedId, err := oauth.ValidateResetToken(ctx, bearerJwt)
+	require.ErrorIs(t, oauth.ErrInvalidToken, err)
+	require.Empty(t, parsedId)
+}
+
+func TestResetTokenIsRSAMethodWithPrivateKey(t *testing.T) {
+	ctx := newContextWithPK()
+	identityId := ksuid.New()
+
+	privateKey, err := rsa.GenerateKey(rand.Reader, 2048)
+	require.NoError(t, err)
+	ctx = runtimectx.WithPrivateKey(ctx, privateKey)
+	require.NoError(t, err)
+
+	jwtToken, err := oauth.GenerateResetToken(ctx, identityId.String())
+	require.NoError(t, err)
+	require.NotEmpty(t, jwtToken)
+
+	_, err = jwt.ParseWithClaims(jwtToken, &oauth.AccessTokenClaims{}, func(token *jwt.Token) (interface{}, error) {
+		if _, ok := token.Method.(*jwt.SigningMethodRSA); !ok {
+			assert.Fail(t, "Invalid signing method. Expected RSA.")
+		}
+		return &privateKey.PublicKey, nil
+	})
+	require.NoError(t, err)
+}
+
+func TestResetTokenHasExpiryClaims(t *testing.T) {
+	ctx := newContextWithPK()
+	identityId := ksuid.New()
+
+	privateKey, err := rsa.GenerateKey(rand.Reader, 2048)
+	require.NoError(t, err)
+	ctx = runtimectx.WithPrivateKey(ctx, privateKey)
+	require.NoError(t, err)
+
+	jwtToken, err := oauth.GenerateResetToken(ctx, identityId.String())
+	require.NoError(t, err)
+	require.NotEmpty(t, jwtToken)
+
+	claims := &oauth.AccessTokenClaims{}
+	_, err = jwt.ParseWithClaims(jwtToken, claims, func(token *jwt.Token) (interface{}, error) {
+		return &privateKey.PublicKey, nil
+	})
+	require.NoError(t, err)
+
+	issuedAt := claims.IssuedAt.Time
+	expiry := claims.ExpiresAt.Time
+
+	require.Greater(t, expiry, time.Now().UTC())
+	require.Equal(t, issuedAt.Add(oauth.ResetTokenExpiry), expiry)
+}
+
+func TestExpiredResetTokenIsInvalid(t *testing.T) {
+	ctx := newContextWithPK()
+	identityId := ksuid.New()
+
+	privateKey, err := rsa.GenerateKey(rand.Reader, 2048)
+	require.NoError(t, err)
+	ctx = runtimectx.WithPrivateKey(ctx, privateKey)
+	require.NoError(t, err)
+
+	// Create the jwt 1 second expired.
+	now := time.Now().UTC().Add(-oauth.ResetTokenExpiry).Add(time.Second * -1)
+	claims := oauth.AccessTokenClaims{
+		RegisteredClaims: jwt.RegisteredClaims{
+			Subject:   identityId.String(),
+			Audience:  jwt.ClaimStrings{"password-reset"},
+			ExpiresAt: jwt.NewNumericDate(now.Add(oauth.ResetTokenExpiry)),
+			IssuedAt:  jwt.NewNumericDate(now),
+		},
+	}
+
+	token := jwt.NewWithClaims(jwt.SigningMethodRS256, claims)
+	tokenString, err := token.SignedString(privateKey)
+	require.NoError(t, err)
+
+	parsedId, err := oauth.ValidateResetToken(ctx, tokenString)
+	require.ErrorIs(t, oauth.ErrTokenExpired, err)
+	require.Empty(t, parsedId)
+}
+
+func TestResetTokenMissingAudIsInvalid(t *testing.T) {
+	ctx := newContextWithPK()
+	identityId := ksuid.New()
+
+	privateKey, err := rsa.GenerateKey(rand.Reader, 2048)
+	require.NoError(t, err)
+	ctx = runtimectx.WithPrivateKey(ctx, privateKey)
+	require.NoError(t, err)
+
+	// Create the jwt with missing aud claim.
+	now := time.Now().UTC()
+	claims := oauth.AccessTokenClaims{
+		RegisteredClaims: jwt.RegisteredClaims{
+			Subject:   identityId.String(),
+			ExpiresAt: jwt.NewNumericDate(now.Add(oauth.ResetTokenExpiry)),
+			IssuedAt:  jwt.NewNumericDate(now),
+		},
+	}
+
+	token := jwt.NewWithClaims(jwt.SigningMethodRS256, claims)
+	tokenString, err := token.SignedString(privateKey)
+	require.NoError(t, err)
+
+	parsedId, err := oauth.ValidateResetToken(ctx, tokenString)
+	require.ErrorIs(t, oauth.ErrInvalidToken, err)
+	require.Empty(t, parsedId)
 }
